pyproject.toml
[project]
name = "torgi-opendata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "pyodbc>=5.0.0",
]
